Fix test regarding different newlines
Commit c206bc869, which updated from the source-map fork to source-map
0.1.38, should have broken the tests, since the fork supported ancient
`'\r'` newlines, while the official package does not. That meant that
the test regarding newlines was no good. Therefore it has been remove.
Instead, another test was updated. We need to both check that the
newlines in the output aren't normalized or so, and that both unix and
Windows newlines increment the line number in the source map.